+++ utils/security/index.ts
@@ -1,2 +1,3 @@
 export { buildContentSecurityPolicy } from './buildContentSecurityPolicy'
 export { buildSecurityHeaders } from './buildSecurityHeaders'
+export { getAllowedOriginsFromEnv } from './getAllowedOriginsFromEnv'